pyproject.toml
[project]
name = "langfuse-eg"
version = "0.1.0"
description = ""
authors = [
    {name = "evangriffiths",email = "evan.griffiths@grantify.io"}
]
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "openai (>=1.63.2,<2.0.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "langfuse (>=2.59.3,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
